Hotfix: window class
tkt.Button() is a bit hard to work with.. I wanted to make a class for some personal practice and it was somewhat reasonable to create a class for the window, but it ended up backfiring b/c  the tkt.Button() method requires a command input which only takes functions. If I really wanted to, I guess I can make a separate file for the goBack function so I can import it into newWindow to avoid dependencies between main and the newWindow class.